Visualization - Pandas
Collection of programs with introduction to the powerful module of pandas. We reviewed many features of pandas to be used in our final assignment. I personally saw the benefit of pandas when it came to processing a csv file. A data exploration using a for loop instead of  pandas was considerably slower. I am exploring many other features outside of the course in projects. In this repo find some sample CSV files for practice.