Numpy
Collection of programs demonstrating the basic use of numpy.
We used numpy as a numerical processing module. In our team exercises, we analysed the data from a data source containing uefa football data. We processed this with a pre-context that players who begin playing football at an earlier age have stats that are higher than others. This was visible by using matplot and outputing to a graph. Overall another great module and good ending to the semester.